+++ src/GitHub_Updater/Language_Pack.php
@@ -112,4 +112,5 @@ class Language_Pack extends Base {
 
 		return $transient;
 	}
+
 }