+++ src/GitHub_Updater/Settings.php
@@ -68,16 +68,67 @@ class Settings extends Base {
 	 * By defining in a method, strings can be translated.
 	 *
 	 * @access private
-	 *
 	 * @return array
 	 */
-	private function _settings_tabs() {
+	private function settings_tabs() {
 		return array(
 			'github_updater_settings'          => esc_html__( 'Settings', 'github-updater' ),
 			'github_updater_install_plugin'    => esc_html__( 'Install Plugin', 'github-updater' ),
 			'github_updater_install_theme'     => esc_html__( 'Install Theme', 'github-updater' ),
 			'github_updater_remote_management' => esc_html__( 'Remote Management', 'github-updater' ),
 		);
+	}
+
+	/**
+	 * Set up the Settings Sub-tabs.
+	 *
+	 * @access private
+	 * @return array
+	 */
+	private function settings_sub_tabs() {
+		$subtabs          = array();
+		$gits             = $this->installed_git_repos();
+		$default_subtabs  = array(
+			'github_updater' => esc_html__( 'GitHub Updater', 'github-updater' ),
+			'github'         => esc_html__( 'GitHub', 'github-updater' ),
+		);
+		$bitbucket_subtab = array( 'bitbucket' => esc_html__( 'Bitbucket', 'github-updater' ) );
+		$gitlab_subtab    = array( 'gitlab' => esc_html__( 'GitLab', 'github-updater' ) );
+		if ( in_array( 'bitbucket', $gits ) ) {
+			$subtabs = array_merge( $subtabs, $bitbucket_subtab );
+		}
+		if ( in_array( 'gitlab', $gits ) ) {
+			$subtabs = array_merge( $subtabs, $gitlab_subtab );
+		}
+
+		return array_merge( $default_subtabs, $subtabs );
+	}
+
+	/**
+	 * Return an array of the installed repository types.
+	 *
+	 * @access private
+	 * @return array
+	 */
+	private function installed_git_repos() {
+		$plugins = Plugin::instance()->get_plugin_configs();
+		$themes  = Theme::instance()->get_theme_configs();
+
+		$repos = array_merge( $plugins, $themes );
+		$gits  = array_map( function( $e ) {
+			return $e->type;
+		}, $repos );
+
+		$gits = array_unique( array_values( $gits ) );
+
+		$gits = array_map( function( $e ) {
+			$e = explode( '_', $e );
+
+			return $e[0];
+		}, $gits );
+
+
+		return array_unique( $gits );
 	}
 
 	/**
@@ -112,14 +163,30 @@ class Settings extends Base {
 	 *
 	 * @access private
 	 */
-	private function _options_tabs() {
+	private function options_tabs() {
 		$current_tab = isset( $_GET['tab'] ) ? $_GET['tab'] : 'github_updater_settings';
 		echo '<h2 class="nav-tab-wrapper">';
-		foreach ( $this->_settings_tabs() as $key => $name ) {
+		foreach ( $this->settings_tabs() as $key => $name ) {
 			$active = ( $current_tab == $key ) ? 'nav-tab-active' : '';
 			echo '<a class="nav-tab ' . $active . '" href="?page=github-updater&tab=' . $key . '">' . $name . '</a>';
 		}
 		echo '</h2>';
+	}
+
+	/**
+	 * Render the settings sub-tabs.
+	 *
+	 * @access private
+	 */
+	private function options_sub_tabs() {
+		$current_tab = isset( $_GET['subtab'] ) ? $_GET['subtab'] : 'github_updater';
+		echo '<h3 class="nav-tab-wrapper">';
+		foreach ( $this->settings_sub_tabs() as $key => $name ) {
+			$active = ( $current_tab == $key ) ? 'nav-tab-active' : '';
+			echo '<a class="nav-tab ' . $active . '" href="?page=github-updater&tab=github_updater_settings&subtab=' . $key . '">' . $name . '</a>';
+		}
+		echo '</h3>';
+
 	}
 
 	/**
@@ -128,6 +195,7 @@ class Settings extends Base {
 	public function create_admin_page() {
 		$action = is_multisite() ? 'edit.php?action=github-updater' : 'options.php';
 		$tab    = isset( $_GET['tab'] ) ? $_GET['tab'] : 'github_updater_settings';
+		$subtab = isset( $_GET['subtab'] ) ? $_GET['subtab'] : 'github_updater';
 		$logo   = plugins_url( basename( dirname( dirname( __DIR__ ) ) ) . '/assets/GitHub_Updater_logo_small.png' );
 		?>
 		<div class="wrap">
@@ -135,7 +203,7 @@ class Settings extends Base {
 				<a href="https://github.com/afragen/github-updater" target="_blank"><img src="<?php esc_attr_e( $logo ); ?>" alt="GitHub Updater logo" /></a><br>
 				<?php esc_html_e( 'GitHub Updater', 'github-updater' ); ?>
 			</h2>
-			<?php $this->_options_tabs(); ?>
+			<?php $this->options_tabs(); ?>
 			<?php if ( ! isset( $_GET['settings-updated'] ) ): ?>
 				<?php if ( is_multisite() && ( isset( $_GET['updated'] ) && true == $_GET['updated'] ) ): ?>
 					<div class="updated">
@@ -152,17 +220,53 @@ class Settings extends Base {
 				<?php endif; ?>
 
 				<?php if ( 'github_updater_settings' === $tab ) : ?>
-					<?php $refresh_transients = add_query_arg( array( 'github_updater_refresh_transients' => true ), $action ); ?>
-					<form method="post" action="<?php esc_attr_e( $refresh_transients ); ?>">
-						<?php submit_button( esc_html__( 'Refresh Transients', 'github-updater' ) ); ?>
-					</form>
-
-					<form method="post" action="<?php esc_attr_e( $action ); ?>">
+					<?php $this->options_sub_tabs(); ?>
+					<form class="settings" method="post" action="<?php esc_attr_e( $action ); ?>">
 						<?php
 						settings_fields( 'github_updater' );
-						do_settings_sections( 'github_updater_install_settings' );
+						switch ( $subtab ) {
+							case 'github_updater':
+								do_settings_sections( 'github_updater_install_settings' );
+								echo '<div style="display:none;">';
+								do_settings_sections( 'github_updater_github_install_settings' );
+								do_settings_sections( 'github_updater_bitbucket_install_settings' );
+								do_settings_sections( 'github_updater_gitlab_install_settings' );
+								echo '</div>';
+								break;
+							case 'github':
+								do_settings_sections( 'github_updater_github_install_settings' );
+								$this->display_ghu_repos( 'github' );
+								echo '<div style="display:none;">';
+								do_settings_sections( 'github_updater_install_settings' );
+								do_settings_sections( 'github_updater_bitbucket_install_settings' );
+								do_settings_sections( 'github_updater_gitlab_install_settings' );
+								echo '</div>';
+								break;
+							case 'bitbucket':
+								do_settings_sections( 'github_updater_bitbucket_install_settings' );
+								$this->display_ghu_repos( 'bitbucket' );
+								echo '<div style="display:none;">';
+								do_settings_sections( 'github_updater_install_settings' );
+								do_settings_sections( 'github_updater_github_install_settings' );
+								do_settings_sections( 'github_updater_gitlab_install_settings' );
+								echo '</div>';
+								break;
+							case 'gitlab':
+								do_settings_sections( 'github_updater_gitlab_install_settings' );
+								$this->display_ghu_repos( 'gitlab' );
+								echo '<div style="display:none;">';
+								do_settings_sections( 'github_updater_install_settings' );
+								do_settings_sections( 'github_updater_github_install_settings' );
+								do_settings_sections( 'github_updater_bitbucket_install_settings' );
+								echo '</div>';
+								break;
+						}
 						submit_button();
 						?>
+					</form>
+					<?php $refresh_transients = add_query_arg( array( 'github_updater_refresh_transients' => true ), $action ); ?>
+					<form class="settings" method="post" action="<?php esc_attr_e( $refresh_transients ); ?>">
+						<?php submit_button( esc_html__( 'Refresh Transients', 'github-updater' ) ); ?>
 					</form>
 				<?php endif; ?>
 			<?php endif; ?>
@@ -177,16 +281,17 @@ class Settings extends Base {
 			?>
 			<?php if ( 'github_updater_remote_management' === $tab ) : ?>
 				<?php $action = add_query_arg( 'tab', $tab, $action ); ?>
-				<?php $reset_api_action = add_query_arg( array( 'github_updater_reset_api_key' => true ), $action ); ?>
-				<form method="post" action="<?php esc_attr_e( $reset_api_action ); ?>">
-					<?php submit_button( esc_html__( 'Reset RESTful key', 'github-updater' ) ); ?>
-				</form>
-				<form method="post" action="<?php esc_attr_e( $action ); ?>">
+
+				<form class="settings" method="post" action="<?php esc_attr_e( $action ); ?>">
 					<?php
 					settings_fields( 'github_updater_remote_management' );
 					do_settings_sections( 'github_updater_remote_settings' );
 					submit_button();
 					?>
+				</form>
+				<?php $reset_api_action = add_query_arg( array( 'github_updater_reset_api_key' => true ), $action ); ?>
+				<form class="settings no-sub-tabs" method="post" action="<?php esc_attr_e( $reset_api_action ); ?>">
+					<?php submit_button( esc_html__( 'Reset RESTful key', 'github-updater' ) ); ?>
 				</form>
 			<?php endif; ?>
 		</div>
@@ -198,6 +303,10 @@ class Settings extends Base {
 	 * Check to see if it's a private repo.
 	 */
 	public function page_init() {
+
+		if ( $this->is_doing_ajax() ) {
+			return;
+		}
 
 		register_setting(
 			'github_updater',           // Option group
@@ -233,14 +342,14 @@ class Settings extends Base {
 			'github_access_token',
 			esc_html__( 'Personal GitHub Access Token', 'github-updater' ),
 			array( &$this, 'print_section_github_access_token' ),
-			'github_updater_install_settings'
+			'github_updater_github_install_settings'
 		);
 
 		add_settings_field(
 			'github_access_token',
 			esc_html__( 'GitHub.com Access Token', 'github-updater' ),
 			array( &$this, 'token_callback_text' ),
-			'github_updater_install_settings',
+			'github_updater_github_install_settings',
 			'github_access_token',
 			array( 'id' => 'github_access_token', 'token' => true )
 		);
@@ -250,7 +359,7 @@ class Settings extends Base {
 				'github_enterprise_token',
 				esc_html__( 'GitHub Enterprise Access Token', 'github-updater' ),
 				array( &$this, 'token_callback_text' ),
-				'github_updater_install_settings',
+				'github_updater_github_install_settings',
 				'github_access_token',
 				array( 'id' => 'github_enterprise_token', 'token' => true )
 			);
@@ -264,7 +373,7 @@ class Settings extends Base {
 				'github_id',
 				esc_html__( 'GitHub Private Settings', 'github-updater' ),
 				array( &$this, 'print_section_github_info' ),
-				'github_updater_install_settings'
+				'github_updater_github_install_settings'
 			);
 		}
 
@@ -277,7 +386,7 @@ class Settings extends Base {
 				'gitlab_settings',
 				esc_html__( 'GitLab Private Settings', 'github-updater' ),
 				array( &$this, 'print_section_gitlab_token' ),
-				'github_updater_install_settings'
+				'github_updater_gitlab_install_settings'
 			);
 		}
 
@@ -286,7 +395,7 @@ class Settings extends Base {
 				'gitlab_private_token',
 				esc_html__( 'GitLab.com Private Token', 'github-updater' ),
 				array( &$this, 'token_callback_text' ),
-				'github_updater_install_settings',
+				'github_updater_gitlab_install_settings',
 				'gitlab_settings',
 				array( 'id' => 'gitlab_private_token', 'token' => true )
 			);
@@ -297,7 +406,7 @@ class Settings extends Base {
 				'gitlab_enterprise_token',
 				esc_html__( 'GitLab CE or GitLab Enterprise Private Token', 'github-updater' ),
 				array( &$this, 'token_callback_text' ),
-				'github_updater_install_settings',
+				'github_updater_gitlab_install_settings',
 				'gitlab_settings',
 				array( 'id' => 'gitlab_enterprise_token', 'token' => true )
 			);
@@ -310,14 +419,14 @@ class Settings extends Base {
 			'bitbucket_user',
 			esc_html__( 'Bitbucket Private Settings', 'github-updater' ),
 			array( &$this, 'print_section_bitbucket_username' ),
-			'github_updater_install_settings'
+			'github_updater_bitbucket_install_settings'
 		);
 
 		add_settings_field(
 			'bitbucket_username',
 			esc_html__( 'Bitbucket Username', 'github-updater' ),
 			array( &$this, 'token_callback_text' ),
-			'github_updater_install_settings',
+			'github_updater_bitbucket_install_settings',
 			'bitbucket_user',
 			array( 'id' => 'bitbucket_username' )
 		);
@@ -326,7 +435,7 @@ class Settings extends Base {
 			'bitbucket_password',
 			esc_html__( 'Bitbucket Password', 'github-updater' ),
 			array( &$this, 'token_callback_text' ),
-			'github_updater_install_settings',
+			'github_updater_bitbucket_install_settings',
 			'bitbucket_user',
 			array( 'id' => 'bitbucket_password', 'token' => true )
 		);
@@ -339,21 +448,30 @@ class Settings extends Base {
 				'bitbucket_id',
 				esc_html__( 'Bitbucket Private Repositories', 'github-updater' ),
 				array( &$this, 'print_section_bitbucket_info' ),
-				'github_updater_install_settings'
+				'github_updater_bitbucket_install_settings'
 			);
 		}
 
 		/*
 		 * Show if no private repositories are present.
 		 */
-		if ( ! parent::$auth_required['github_private'] && ! parent::$auth_required['bitbucket_private'] ) {
+		if ( ! parent::$auth_required['github_private'] ) {
 			add_settings_section(
 				null,
-				esc_html__( 'No private repositories are installed.', 'github-updater' ),
+				esc_html__( 'Private repositories are not installed, are cached, or are using your personal access token.', 'github-updater' ),
 				array(),
-				'github_updater_install_settings'
+				'github_updater_github_install_settings'
 			);
 		}
+		if ( ! parent::$auth_required['bitbucket_private'] ) {
+			add_settings_section(
+				null,
+				esc_html__( 'Private repositories are not installed or are cached.', 'github-updater' ),
+				array(),
+				'github_updater_bitbucket_install_settings'
+			);
+		}
+
 
 		$this->update_settings();
 	}
@@ -434,21 +552,23 @@ class Settings extends Base {
 
 			$setting_field['id']    = $token->repo;
 			$setting_field['title'] = $type . $token->name;
-			$setting_field['page']  = 'github_updater_install_settings';
 
 			$token_type = explode( '_', $token->type );
 			switch ( $token_type[0] ) {
 				case 'github':
+					$setting_field['page']            = 'github_updater_github_install_settings';
 					$setting_field['section']         = 'github_id';
 					$setting_field['callback_method'] = array( &$this, 'token_callback_text' );
 					$setting_field['callback']        = $token->repo;
 					break;
 				case 'bitbucket':
+					$setting_field['page']            = 'github_updater_bitbucket_install_settings';
 					$setting_field['section']         = 'bitbucket_id';
 					$setting_field['callback_method'] = array( &$this, 'token_callback_checkbox' );
 					$setting_field['callback']        = $token->repo;
 					break;
 				case 'gitlab':
+					$setting_field['page']            = 'github_updater_gitlab_install_settings';
 					$setting_field['section']         = 'gitlab_id';
 					$setting_field['callback_method'] = array( &$this, 'token_callback_checkbox' );
 					$setting_field['callback']        = $token->repo;
@@ -470,8 +590,9 @@ class Settings extends Base {
 		 */
 		$ghu_unset_keys = array_diff_key( parent::$options, $ghu_options_keys );
 		$always_unset   = array(
-			'github_access_token',
 			'branch_switch',
+			'github_access_token',
+			'github_enterprise_token',
 			'bitbucket_username',
 			'bitbucket_password',
 		);
@@ -635,7 +756,7 @@ class Settings extends Base {
 		$type = ( isset( $args['token'] ) ) ? 'password' : 'text';
 		?>
 		<label for="<?php esc_attr( $args['id'] ); ?>">
-			<input type="<?php esc_attr_e( $type ); ?>" style="width:50%;" name="github_updater[<?php esc_attr_e( $args['id'] ); ?>]" value="<?php esc_attr_e( $name ); ?>">
+			<input class="ghu-callback-text" type="<?php esc_attr_e( $type ); ?>" name="github_updater[<?php esc_attr_e( $args['id'] ); ?>]" value="<?php esc_attr_e( $name ); ?>">
 		</label>
 		<?php
 	}
@@ -680,13 +801,50 @@ class Settings extends Base {
 	public function update_settings() {
 		if ( isset( $_POST['option_page'] ) ) {
 			if ( 'github_updater' === $_POST['option_page'] ) {
-				update_site_option( 'github_updater', self::sanitize( $_POST['github_updater'] ) );
+				$options = $this->filter_options();
+				update_site_option( 'github_updater', self::sanitize( $options ) );
 			}
 			if ( 'github_updater_remote_management' === $_POST['option_page'] ) {
 				update_site_option( 'github_updater_remote_management', (array) self::sanitize( $_POST['github_updater_remote_management'] ) );
 			}
 		}
 		$this->redirect_on_save();
+	}
+
+	/**
+	 * Filter options so that sub-tab options are grouped in single $options variable.
+	 *
+	 * @access private
+	 * @return array|mixed
+	 */
+	private function filter_options() {
+		$plugins          = Plugin::instance()->get_plugin_configs();
+		$themes           = Theme::instance()->get_theme_configs();
+		$repos            = array_merge( $plugins, $themes );
+		$options          = parent::$options;
+		$non_repo_options = array(
+			'github_access_token',
+			'bitbucket_username',
+			'bitbucket_password',
+			'gitlab_private_token',
+			'gitlab_enterprise_token',
+			'branch_switch',
+		);
+
+		$repos = array_map( function( $e ) {
+			return $e->repo = null;
+		}, $repos );
+
+		array_filter( $non_repo_options,
+			function( $e ) use ( &$options ) {
+				unset( $options[ $e ] );
+			}
+		);
+
+		$intersect = array_intersect( $options, $repos );
+		$options   = array_merge( $intersect, $_POST['github_updater'] );
+
+		return $options;
 	}
 
 	/**
@@ -710,12 +868,14 @@ class Settings extends Base {
 		if ( $update || $refresh_transients || $reset_api_key ) {
 			$query = isset( $_POST['_wp_http_referer'] ) ? parse_url( $_POST['_wp_http_referer'], PHP_URL_QUERY ) : null;
 			parse_str( $query, $arr );
-			$arr['tab'] = ! empty( $arr['tab'] ) ? $arr['tab'] : 'github_updater_settings';
+			$arr['tab']    = ! empty( $arr['tab'] ) ? $arr['tab'] : 'github_updater_settings';
+			$arr['subtab'] = ! empty( $arr['subtab'] ) ? $arr['subtab'] : 'github_updater';
 
 			$location = add_query_arg(
 				array(
 					'page'               => 'github-updater',
 					'tab'                => $arr['tab'],
+					'subtab'             => $arr['subtab'],
 					'refresh_transients' => $refresh_transients,
 					'reset'              => $reset_api_key,
 					'updated'            => $update,
@@ -777,6 +937,38 @@ class Settings extends Base {
 		$link          = array( '<a href="' . esc_url( network_admin_url( $settings_page ) ) . '?page=github-updater">' . esc_html__( 'Settings', 'github-updater' ) . '</a>' );
 
 		return array_merge( $links, $link );
+	}
+
+	/**
+	 * Write out listing of installed plugins and themes using GitHub Updater.
+	 *
+	 * @param $type
+	 */
+	private function display_ghu_repos( $type ) {
+		$plugins = Plugin::instance()->get_plugin_configs();
+		$themes  = Theme::instance()->get_theme_configs();
+		$repos   = array_merge( $plugins, $themes );
+
+		$type_repos = array_filter( $repos, function( $e ) use ( $type ) {
+			return false !== stristr( $e->type, $type );
+		} );
+
+		$display_data = array_map( function( $e ) {
+			return $e = array(
+				'type' => $e->type,
+				'repo' => $e->repo,
+				'name' => $e->name,
+			);
+		}, $type_repos );
+
+		printf( '<h4>' . esc_html__( 'Installed Plugins and Themes', 'github-updater' ) . '</h4>' );
+		foreach ( $display_data as $data ) {
+			$dashicon = '<span class="dashicons dashicons-admin-plugins"></span>&nbsp;';
+			if ( false !== strpos( $data['type'], 'theme' ) ) {
+				$dashicon = '<span class="dashicons dashicons-admin-appearance"></span>&nbsp;';
+			}
+			printf( '<p>' . $dashicon . ' ' . $data['name'] . '</p>' );
+		}
 	}
 
 }